% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Title My Binomial Distribution}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of iterations}

\item{n}{Number of trials}

\item{p}{Probability of trial success}
}
\value{
An algorithm that makes a barplot of expected values from a binomial distribution
}
\description{
Title My Binomial Distribution
}
\examples{
\dontrun{mybin(100,10,.5)}
}
