% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{MyCI}
\usage{
myci(x)
}
\arguments{
\item{x}{sample data}
}
\value{
A 95% confidence interval for the mean of the population being sampled
}
\description{
MyCI
}
\examples{
\dontrun{myci(x = rnorm(30, mean = 10, sd = 12))}
}
