% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{mycltu}
\alias{mycltu}
\title{MyCLT}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{number of trials per iteration}

\item{iter}{number of iterations}

\item{a}{minimum value of sample}

\item{b}{maximum value of sample}
}
\value{
a histogram with the sample means of repeated sampling
}
\description{
MyCLT
}
\examples{
\dontrun{mycltu(n = 10, iter = 10000, a = 0, b = 10)}
}
