% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{MyDDT}
\usage{
myddt(df, SPECIES)
}
\arguments{
\item{df}{the data frame}

\item{SPECIES}{desired species}
}
\value{
a plot, a csv file, and 3 command-line outputs
}
\description{
Project 1: myddt function
}
\examples{
\dontrun{myddt(df = ddt, species = "CCATFISH")}

}
