% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{MyMLNorm}
\usage{
mymlnorm(x, mu, sig)
}
\arguments{
\item{x}{data values vector}

\item{mu}{range of possible values for the mean}

\item{sig}{range of possible values for the standard deviation}
}
\value{
A contour plot highlighting the mean and standard deviation for the given data values vector, assuming that the dataset is normally distributed
}
\description{
MyMLNorm
}
\examples{
\dontrun{mymlnorm(x=c(10,12,13,15,12,11),mu=seq(10,15,length=1000),sig=seq(0.1,4,length=1000))}
}
