% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Title MyNCurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of distribution}

\item{sigma}{Std. Deviation of Distribution}

\item{a}{Requested value}
}
\value{

}
\description{
Title MyNCurve
}
