% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{My P Value}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{Calculated t-statistic for the sample being tested}

\item{xmax}{Upper and lower bounds of the graph (default 4)}

\item{n}{Number of experimental units}

\item{alpha}{Significance  level}
}
\value{
A plot of the t-statistic, the rejection region, and the p-value
}
\description{
My P Value
}
\examples{
\dontrun{mypvalue(tcalc,n=30,alpha=0.05)}
}
